factor AOT types out to a `stages` module
